backend/pyproject.toml
[project]
name = "audit-iq-backend"
version = "0.1.0"
description = "Backend FastAPI pour Audit-IQ - Plateforme d'audit d'équité IA"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.1",
    "python-multipart>=0.0.12",
    "pandas>=2.2.3",
    "numpy>=2.2.0",
    "scikit-learn>=1.6.0",
    "fairlearn>=0.11.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.3",
    "email-validator>=2.0.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.27.0",
    "alembic>=1.11.0",
    "pydantic-settings>=2.6.1",
    "bcrypt>=4.2.1",
    "PyJWT>=2.10.0",
    "openpyxl>=3.1.5",
    "aif360>=0.6.1",
    "aiosqlite"
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["audit_iq_backend"]
